Initial WIP refactor
